+++ polyfill/lib/index.mjs
@@ -1,24 +1,10 @@
+// This entry point treats Temporal as a library, and does not polyfill it onto
+// the global object.
+// This is in order to avoid breaking the web in the future, if the polyfill
+// gains wide adoption before the API is finalized. We do not want checks such
+// as `if (typeof Temporal === 'undefined')` in the wild, until browsers start
+// shipping the finalized API.
+
 import * as Temporal from './temporal.mjs';
 import * as Intl from './intl.mjs';
-
-Object.defineProperty(globalThis, 'Temporal', {
-  value: {},
-  writable: true,
-  enumerable: false,
-  configurable: true
-});
-copy(globalThis.Temporal, Temporal);
-copy(globalThis.Intl, Intl);
-
-function copy(target, source) {
-  for (const prop of Object.getOwnPropertyNames(source)) {
-    Object.defineProperty(target, prop, {
-      value: source[prop],
-      writable: true,
-      enumerable: false,
-      configurable: true
-    });
-  }
-}
-
 export { Temporal, Intl };
