+++ polyfill/lib/shim.mjs
@@ -1,0 +1,28 @@
+// This is an alternate entry point that polyfills Temporal onto the global
+// object. This is used only for the browser playground and the test262 tests.
+// See the note in index.mjs.
+
+import * as Temporal from './temporal.mjs';
+import * as Intl from './intl.mjs';
+
+Object.defineProperty(globalThis, 'Temporal', {
+  value: {},
+  writable: true,
+  enumerable: false,
+  configurable: true
+});
+copy(globalThis.Temporal, Temporal);
+copy(globalThis.Intl, Intl);
+
+function copy(target, source) {
+  for (const prop of Object.getOwnPropertyNames(source)) {
+    Object.defineProperty(target, prop, {
+      value: source[prop],
+      writable: true,
+      enumerable: false,
+      configurable: true
+    });
+  }
+}
+
+export { Temporal, Intl };
